# swagger.yml
swagger: "2.0"
info:
  description: "My Hacker News documentation"
  version: "1.0.0"
  title: "My Hacker News"
host: "myhackernews.tk"
basePath: "/api"
...
schemes:
- "http"
paths:
  /asks:
    get:
      tags:
        - "contribution"
      summary: "Endpoint to view the asks"
      description: "Returns all the ask contributions sorted by n_votes"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
            description: "Successful operation. Returns the array of asks"
            schema:
              type: "array"
              items:
                $ref: "#/definitions/Ask"
    post:
      tags:
        - "contribution"
      summary: "Endpoint to create new ask contributions"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Ask"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/New_Ask"
      responses:
        201:
            description: "Successful operation. Returns created contribution"
            schema:
              $ref: "#/definitions/Ask"
        400:
            description: "Bad request, there are missing parameters or they do not have valid format"
        401:
            description: "Forbidden, missing authentication"
      security:
      - api_key: []
  /news:
    get:
      tags:
        - "contribution"
      summary: "Endpoint to view the news"
      description: "Returns all the news contributions sorted by n_votes"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
            description: "Successful operation. Returns the array of news"
            schema:
              type: "array"
              items:
                $ref: "#/definitions/New"
    post:
      tags:
        - "contribution"
      summary: "Endpoint to create new new contributions"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "New"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/New_New"
      responses:
        201:
            description: "Successful operation. Returns created contribution"
            schema:
              $ref: "#/definitions/New"
        400:
            description: "Bad request, there are missing parameters or they do not have valid format"
        401:
            description: "Forbidden, missing authentication"
        409:
            description: "Conflict, specified URL already posted"
            schema:
              $ref: "#/definitions/New"
      security:
      - api_key: []
...
  /threads/{userID}:
    get:
      tags:
        - "comment"
      summary: "Endpoint to get the comments of an specific user"
      description: "Returns all the comments of the user"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "path"
          name: "userID"
          description: "ID of the user the comments of which we want to return"
          required: true
          type: "string"
      responses:
        200:
            description: "Successful operation. Returns the array of comments"
            schema:
              type: "array"
              items:
                $ref: "#/definitions/Comment"
        404:
            description: "User not found"
  /contributions/{contributionID}:
    get:
      tags:
        - "contribution"
      summary: "Endpoint to get an specific contribution"
      description: "Returns a single contribution"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "path"
          name: "contributionID"
          description: "ID of contribution to return"
          required: true
          type: "string"
      responses:
        200:
            description: "Successful operation. Returns asked contribution"
            schema:
              $ref: "#/definitions/Contribution"
        400:
            description: "Bad request, there are missing parameters or they do not have valid format"
        404:
            description: "Contribution not found"
    delete:
      tags:
        - "contribution"
      summary: "Endpoint to delete an especific contribution"
      description: "Deletes a contribution and all of it's comments"
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - in: "path"
          name: "contributionID"
          description: "ID of the contribution to delete"
          required: true
          type: "string"
      responses:
        200:
            description: "Suuccesscul operation. Contribution deleted"
        400:
            description: "Bad request, there are missing parameteres or they do not have a valid format"
        401:
            description: "Forbidden. Missing Authentication"
        403:
            description: "Unauthorized. Missing permissions to delete the contribution"
        404:
            description: "Contribution not found"
  /contributions/{contributionID}/vote:
    post:
      tags:
        - "contribution"
      summary: "Endpoint to vote an specific contribution"
      description: "Returns a single contribution"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "path"
          name: "contributionID"
          description: "ID of contribution to return"
          required: true
          type: "string"
      responses:
        200:
            description: "Successful operation. Returns voted contribution"
            schema:
              $ref: "#/definitions/Contribution"
        400:
            description: "Bad request, there are missing parameters or they do not have valid format"
        401:
            description: "Forbidden, missing authentication"
        403:
            description: "Unauthorized. Missing permissions to update the resources"
        404:
            description: "Contribution not found"
        409:
            description: "Conflict. You have already vote for this contribution"
      security:
      - api_key: []
    delete:
      tags:
        - "contribution"
      summary: "Endpoint to unvote an specific contribution"
      description: "Returns a single contribution"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "path"
          name: "contributionID"
          description: "ID of contribution to return"
          required: true
          type: "string"
      responses:
        200:
            description: "Successful operation. Returns unvoted contribution"
            schema:
              $ref: "#/definitions/Contribution"
        400:
            description: "Bad request, there are missing parameters or they do not have valid format"
        401:
            description: "Forbidden, missing authentication"
        403:
            description: "Unauthorized. Missing permissions to update the resources"
        404:
            description: "Contribution not found"
        409:
            description: "Conflict. You have already vote for this contribution"
      security:
      - api_key: []
  /comments/{commentID}:
    get:
      tags:
        - "comment"
      summary: "Endpoint to get an specific comment"
      description: "Returns a single comment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "path"
          name: "commentID"
          description: "ID of comment to return"
          required: true
          type: "string"
      responses:
        200:
            description: "Successful operation. Returns asked comment"
            schema:
              $ref: "#/definitions/Comment"
        400:
            description: "Bad request, there are missing parameters or they do not have valid format"
        404:
            description: "Comment not found"
    delete:
      tags:
        - "comment"
      summary: "Endpoint to delete an specific comment"
      description: "Deletes an specific comment and all of it's sons"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "commentID"
          description: "ID of the comment to delete"
          required: true
          type: "string"
      responses:
        200:
            description: "Suuccesscul operation. Comment deleted"
        400:
            description: "Bad request, there are missing parameteres or they do not have a valid format"
        401:
            description: "Forbidden. Missing Authentication"
        403:
            description: "Unauthorized. Missing permissions to delete the comment"
        404:
            description: "Comment not found"
  /comments/{commentID}/vote:
    post:
      tags:
        - "comment"
      summary: "Endpoint to vote an specific comment"
      description: "Returns a single comment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "path"
          name: "commentID"
          description: "ID of a comment to return"
          required: true
          type: "string"
      responses:
        200:
            description: "Successful operation. Returns voted comment"
            schema:
              $ref: "#/definitions/Comment"
        400:
            description: "Bad request, there are missing parameters or they do not have valid format"
        401:
            description: "Forbidden, missing authentication"
        403:
            description: "Unauthorized. Missing permissions to update the resources"
        404:
            description: "Comment not found"
        409:
            description: "Conflict. You have already vote for this comment"
      security:
      - api_key: []
    delete:
      tags:
        - "comment"
      summary: "Endpoint to unvote an specific comment"
      description: "Returns a single comment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: "path"
          name: "commentID"
          description: "ID of comment to return"
          required: true
          type: "string"
      responses:
        200:
            description: "Successful operation. Returns unvoted comment"
            schema:
              $ref: "#/definitions/Comment"
        400:
            description: "Bad request, there are missing parameters or they do not have valid format"
        401:
            description: "Forbidden, missing authentication"
        403:
            description: "Unauthorized. Missing permissions to update the resources"
        404:
            description: "Comment not found"
        409:
            description: "Conflict. You have already vote for this comment"
      security:
      - api_key: []
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  Contribution:
    type: "object"
    required:
      - "title"
      - "text"
      - "username"
    properties:
      id:
        type: "integer"
        format: "int64"
      kind:
        type: "string"
      n_votes:
        type: "integer"
        format: "int64"
      n_comments:
        type: "integer"
        format: "int 64"
      time:
        type: "string"
        format: "date-time"
      title:
        type: "string"
      text:
        type: "string"
      url:
        type: "string"
      user:
        type: "string"
  New:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      url:
        type: "string"
      time:
        type: "string"
        format: "date-time"
      username:
        type: "string"
      n_votes:
        type: "integer"
        format: "int64"
      n_comments:
        type: "integer"
        format: "int64"
  Ask:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      text:
        type: "string"
      time:
        type: "string"
        format: "date-time"
      username:
        type: "string"
      kind:
        type: "string"
      n_votes:
        type: "integer"
        format: "int64"
  New_Ask:
    type: "object"
    properties:
      title:
        type: "string"
      text:
        type: "string"
  New_New:
    type: "object"
    properties:
      title:
        type: "string"
      url:
        type: "string"
  Comment:
    type: "object"
    required:
      - "text"
      - "username"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      time:
        type: "string"
        format: "date-time"
      text:
        type: "string"
      contribution_id:
        type: "integer"
        format: "int64"
      parent_id:
        type: "integer"
        format: "int64"
      n_votes:
        type: "integer"
        format: "int64"